+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 FEATURES:
 
 - Created provider.
+- Created resources:
+  - `fauna_collection` (Collection)